[VIDEO] - signed vs unsigned

+++ dlib/video_io/video_demuxer_impl.h
@@ -221,7 +221,7 @@ namespace dlib
                 _channel_video._resizer_image.resize(f, tmp);
                 array2d<rgb_pixel> frame_image(tmp.frame->height, tmp.frame->width);
                 
-                for (size_t row = 0 ; row < tmp.frame->height ; row++)
+                for (int row = 0 ; row < tmp.frame->height ; row++)
                 {
                     memcpy(frame_image.begin() + row * tmp.frame->width, 
                            tmp.frame->data[0]  + row * tmp.frame->linesize[0],
